+++ src/scheduler.cpp
@@ -34,18 +34,21 @@ namespace ctbot {
 using namespace avr;
 
 Scheduler::Task::Task(const uint16_t id, const uint16_t period, task_func_t&& func, task_func_data_t&& func_data)
-    : id_ { id }, period_ { static_cast<uint16_t>(Timer::ms_to_ticks(period)) }, next_runtime_ { 0 }, active_ { true }, func_ { std::move(func) }, func_data_ { std::move(func_data) } {}
+    : id_ { id }, period_ { static_cast<uint16_t>(Timer::ms_to_ticks(period)) }, next_runtime_ { 0 }, active_ { true }, func_ { std::move(func) }, func_data_ {
+          std::move(func_data)
+      } {}
 
 Scheduler::Task::Task(const uint16_t id, const uint16_t period, const uint32_t next_run, task_func_t&& func, task_func_data_t&& func_data)
-    : id_ { id }, period_ { static_cast<uint16_t>(Timer::ms_to_ticks(period)) }, next_runtime_ { next_run }, active_ { true }, func_ { std::move(func) }, func_data_ { std::move(func_data) } {}
+    : id_ { id }, period_ { static_cast<uint16_t>(Timer::ms_to_ticks(period)) }, next_runtime_ { next_run }, active_ { true }, func_ { std::move(func) },
+      func_data_ { std::move(func_data) } {}
 
-std::ostream& operator <<(std::ostream& os, const Scheduler::Task& v) {
+std::ostream& operator<<(std::ostream& os, const Scheduler::Task& v) {
     os << std::hex << "0x" << v.id_ << "\t" << (v.active_ ? F("  ACTIVE") : F("INACTIVE")) << "\t" << std::dec << Timer::ticks_to_us(v.period_) << " us\n";
     return os;
 }
 
 void Scheduler::run() {
-    while (running_ && (! task_queue_.empty())) {
+    while (running_ && (!task_queue_.empty())) {
         const Task& task { task_queue_.top() };
         // std::cout << "Scheduler::run(): task.next_runtime=" << task.next_runtime_ << " ticks.\n";
         const uint32_t next_runtime { task.next_runtime_ };
@@ -53,7 +56,7 @@ void Scheduler::run() {
         // std::cout << "Scheduler::run(): now=" << now << " ticks \n";
         while (next_runtime > now) {
             do {
-                _SLEEP_CONTROL_REG = (_SLEEP_CONTROL_REG & ~(BV_8(SM0) | BV_8(SM1) | BV_8(SM2))) | (SLEEP_MODE_IDLE);
+                _SLEEP_CONTROL_REG = static_cast<uint8_t>((_SLEEP_CONTROL_REG & ~(BV_8(SM0) | BV_8(SM1) | BV_8(SM2))) | (SLEEP_MODE_IDLE));
             } while (0);
             sleep_enable();
             sleep_cpu();
@@ -77,8 +80,8 @@ uint16_t Scheduler::task_add(const std::string& name, const uint16_t period, tas
     task_queue_.push(task);
     task_names_.push_back(name);
 
-    std::cout << F("Scheduler::task_add(): task \"") << task_names_[next_id_] << F("\" with id 0x") << std::hex << task.id_ << F(" and period of ")
-        << std::dec << Timer::ticks_to_us(task.period_) << F(" us added.\n");
+    std::cout << F("Scheduler::task_add(): task \"") << task_names_[next_id_] << F("\" with id 0x") << std::hex << task.id_ << F(" and period of ") << std::dec
+              << Timer::ticks_to_us(task.period_) << F(" us added.\n");
     return next_id_++;
 }
 
@@ -95,7 +98,7 @@ uint16_t Scheduler::task_get(const std::string& name) const {
 bool Scheduler::task_suspend(const uint16_t id) {
     // std::cout << "Scheduler::task_suspend(0x" << std::hex << id << std::dec << ")\n";
     bool res { false };
-    for (auto& t : task_vector_ ) {
+    for (auto& t : task_vector_) {
         if (t.id_ == id) {
             t.active_ = false;
             res = true;
@@ -109,7 +112,7 @@ bool Scheduler::task_suspend(const uint16_t id) {
 bool Scheduler::task_resume(const uint16_t id) {
     // std::cout << "Scheduler::task_resume(0x" << std::hex << id << std::dec << ")\n";
     bool res { false };
-    for (auto& t : task_vector_ ) {
+    for (auto& t : task_vector_) {
         if (t.id_ == id) {
             t.active_ = true;
             res = true;
